fixed winder regression due to refactor
