+++ server/src/machines/winder2/mod.rs
@@ -2,7 +2,6 @@ pub mod act;
 pub mod adaptive_spool_speed_controller;
 pub mod api;
 pub mod clamp_revolution;
-pub mod emit;
 pub mod filament_tension;
 pub mod minmax_spool_speed_controller;
 pub mod new;
@@ -10,38 +9,52 @@ pub mod puller_speed_controller;
 pub mod spool_speed_controller;
 pub mod tension_arm;
 pub mod traverse_controller;
-
-use api::{SpoolAutomaticActionMode, Winder2Namespace};
+use api::{
+    LiveValuesEvent, ModeState, PullerState, SpoolAutomaticActionMode, SpoolAutomaticActionState,
+    SpoolSpeedControllerState, StateEvent, TensionArmState, TraverseState, Winder2Events,
+    Winder2Namespace,
+};
+use control_core::socketio::event::BuildEvent;
 use control_core::{
     converters::angular_step_converter::AngularStepConverter,
     machines::{
-        ConnectedMachine,
+        ConnectedMachine, ConnectedMachineData, downcast_machine,
         identification::{MachineIdentification, MachineIdentificationUnique},
         manager::MachineManager,
     },
+    socketio::namespace::NamespaceCacheingLogic,
+    uom_extensions::velocity::meter_per_minute,
 };
-
 use control_core_derive::Machine;
 use ethercat_hal::io::{
     digital_input::DigitalInput, digital_output::DigitalOutput,
     stepper_velocity_el70x1::StepperVelocityEL70x1,
 };
-use puller_speed_controller::PullerSpeedController;
+use futures::executor::block_on;
+use puller_speed_controller::{PullerRegulationMode, PullerSpeedController};
 use smol::lock::{Mutex, RwLock};
 use spool_speed_controller::SpoolSpeedController;
-use std::{fmt::Debug, sync::Weak, time::Instant};
+use std::{
+    fmt::Debug,
+    sync::{Arc, Weak},
+    time::Instant,
+};
 use tension_arm::TensionArm;
 use traverse_controller::TraverseController;
 use uom::{
     ConstZero,
     si::{
-        f64::Length,
+        angle::degree,
+        angular_velocity::revolution_per_minute,
+        f64::{Length, Velocity},
         length::{meter, millimeter},
         velocity::meter_per_second,
     },
 };
 
-use crate::machines::{MACHINE_WINDER_V1, VENDOR_QITECH, buffer1::BufferV1};
+use crate::machines::{
+    MACHINE_WINDER_V1, VENDOR_QITECH, buffer1::BufferV1, winder2::api::ConnectedMachineState,
+};
 
 #[derive(Debug)]
 pub struct SpoolAutomaticAction {
@@ -103,14 +116,222 @@ impl Winder2 {
     };
 }
 
-/// Implement Traverse helpers
+/// Implement Traverse
 impl Winder2 {
     fn set_laser(&mut self, value: bool) {
         self.laser.set(value);
+        self.emit_state();
     }
 
+    /// Validates that traverse limits maintain proper constraints:
+    /// - Inner limit must be smaller than outer limit
+    /// - At least 0.9mm difference between inner and outer limits
     fn validate_traverse_limits(inner: Length, outer: Length) -> bool {
         outer > inner + Length::new::<millimeter>(0.9)
+    }
+
+    pub fn traverse_set_limit_inner(&mut self, limit: f64) {
+        let new_inner = Length::new::<millimeter>(limit);
+        let current_outer = self.traverse_controller.get_limit_outer();
+
+        // Validate the new inner limit against current outer limit
+        if !Self::validate_traverse_limits(new_inner, current_outer) {
+            // Don't update if validation fails - keep the current value
+            return;
+        }
+        self.traverse_controller.set_limit_inner(new_inner);
+        self.emit_state();
+    }
+
+    pub fn traverse_set_limit_outer(&mut self, limit: f64) {
+        let new_outer = Length::new::<millimeter>(limit);
+        let current_inner = self.traverse_controller.get_limit_inner();
+
+        // Validate the new outer limit against current inner limit
+        if !Self::validate_traverse_limits(current_inner, new_outer) {
+            // Don't update if validation fails - keep the current value
+            return;
+        }
+
+        self.traverse_controller.set_limit_outer(new_outer);
+        self.emit_state();
+    }
+
+    pub fn traverse_set_step_size(&mut self, step_size: f64) {
+        let step_size = Length::new::<millimeter>(step_size);
+        self.traverse_controller.set_step_size(step_size);
+        self.emit_state();
+    }
+
+    pub fn traverse_set_padding(&mut self, padding: f64) {
+        let padding = Length::new::<millimeter>(padding);
+        self.traverse_controller.set_padding(padding);
+        self.emit_state();
+    }
+
+    pub fn traverse_goto_limit_inner(&mut self) {
+        if self.can_go_in() {
+            self.traverse_controller.goto_limit_inner();
+        }
+        self.emit_state();
+    }
+
+    pub fn traverse_goto_limit_outer(&mut self) {
+        if self.can_go_out() {
+            self.traverse_controller.goto_limit_outer();
+        }
+        self.emit_state();
+    }
+
+    pub fn traverse_goto_home(&mut self) {
+        if self.can_go_home() {
+            self.traverse_controller.goto_home();
+        }
+        self.emit_state();
+    }
+
+    pub fn emit_live_values(&mut self) {
+        let angle_deg = self.tension_arm.get_angle().get::<degree>();
+
+        // Wrap [270;<360] to [-90; 0]
+        // This is done to reduce flicker in the graphs around the zero point
+        let angle_deg = if angle_deg >= 270.0 {
+            angle_deg - 360.0
+        } else {
+            angle_deg
+        };
+
+        // Calculate puller speed from current motor steps
+        let steps_per_second = self.puller.get_speed();
+        let angular_velocity = self
+            .puller_speed_controller
+            .converter
+            .steps_to_angular_velocity(steps_per_second as f64);
+        let puller_speed = self
+            .puller_speed_controller
+            .angular_velocity_to_speed(angular_velocity);
+
+        // Calculate spool RPM from current motor steps
+        let spool_rpm = self
+            .spool_step_converter
+            .steps_to_angular_velocity(self.spool.get_speed() as f64)
+            .get::<revolution_per_minute>();
+
+        let live_values = LiveValuesEvent {
+            traverse_position: self
+                .traverse_controller
+                .get_current_position()
+                .map(|x| x.get::<millimeter>()),
+            puller_speed: puller_speed.get::<meter_per_minute>(),
+            spool_rpm,
+            tension_arm_angle: angle_deg,
+            spool_progress: self.spool_automatic_action.progress.get::<meter>(),
+        };
+
+        let event = live_values.build();
+        self.namespace.emit(Winder2Events::LiveValues(event));
+    }
+
+    pub fn build_state_event(&mut self) -> StateEvent {
+        StateEvent {
+            is_default_state: !std::mem::replace(&mut self.emitted_default_state, true),
+            traverse_state: TraverseState {
+                limit_inner: self
+                    .traverse_controller
+                    .get_limit_inner()
+                    .get::<millimeter>(),
+                limit_outer: self
+                    .traverse_controller
+                    .get_limit_outer()
+                    .get::<millimeter>(),
+                position_in: self
+                    .traverse_controller
+                    .get_limit_inner()
+                    .get::<millimeter>(),
+                position_out: self
+                    .traverse_controller
+                    .get_limit_outer()
+                    .get::<millimeter>(),
+                is_going_in: self.traverse_controller.is_going_in(),
+                is_going_out: self.traverse_controller.is_going_out(),
+                is_homed: self.traverse_controller.is_homed(),
+                is_going_home: self.traverse_controller.is_going_home(),
+                is_traversing: self.traverse_controller.is_traversing(),
+                laserpointer: self.laser.get(),
+                step_size: self.traverse_controller.get_step_size().get::<millimeter>(),
+                padding: self.traverse_controller.get_padding().get::<millimeter>(),
+                can_go_in: self.can_go_in(),
+                can_go_out: self.can_go_out(),
+                can_go_home: self.can_go_home(),
+            },
+            puller_state: PullerState {
+                regulation: self.puller_speed_controller.regulation_mode.clone(),
+                target_speed: self
+                    .puller_speed_controller
+                    .target_speed
+                    .get::<meter_per_minute>(),
+                target_diameter: self
+                    .puller_speed_controller
+                    .target_diameter
+                    .get::<millimeter>(),
+                forward: self.puller_speed_controller.forward,
+            },
+            mode_state: ModeState {
+                mode: self.mode.clone().into(),
+                can_wind: self.can_wind(),
+            },
+            tension_arm_state: TensionArmState {
+                zeroed: self.tension_arm.zeroed,
+            },
+            spool_speed_controller_state: SpoolSpeedControllerState {
+                regulation_mode: self.spool_speed_controller.get_type().clone(),
+                minmax_min_speed: self
+                    .spool_speed_controller
+                    .get_minmax_min_speed()
+                    .get::<revolution_per_minute>(),
+                minmax_max_speed: self
+                    .spool_speed_controller
+                    .get_minmax_max_speed()
+                    .get::<revolution_per_minute>(),
+                adaptive_tension_target: self.spool_speed_controller.get_adaptive_tension_target(),
+                adaptive_radius_learning_rate: self
+                    .spool_speed_controller
+                    .get_adaptive_radius_learning_rate(),
+                adaptive_max_speed_multiplier: self
+                    .spool_speed_controller
+                    .get_adaptive_max_speed_multiplier(),
+                adaptive_acceleration_factor: self
+                    .spool_speed_controller
+                    .get_adaptive_acceleration_factor(),
+                adaptive_deacceleration_urgency_multiplier: self
+                    .spool_speed_controller
+                    .get_adaptive_deacceleration_urgency_multiplier(),
+            },
+            spool_automatic_action_state: SpoolAutomaticActionState {
+                spool_required_meters: self.spool_automatic_action.target_length.get::<meter>(),
+                spool_automatic_action_mode: self.spool_automatic_action.mode.clone(),
+            },
+            connected_machine_state: ConnectedMachineState {
+                machine_identification_unique: self.connected_buffer.as_ref().map(
+                    |connected_machine| {
+                        ConnectedMachineData::from(connected_machine).machine_identification_unique
+                    },
+                ),
+                is_available: self
+                    .connected_buffer
+                    .as_ref()
+                    .map(|connected_machine| {
+                        ConnectedMachineData::from(connected_machine).is_available
+                    })
+                    .unwrap_or(false),
+            },
+        }
+    }
+
+    pub fn emit_state(&mut self) {
+        let state_event = self.build_state_event();
+        let event = state_event.build();
+        self.namespace.emit(Winder2Events::State(event));
     }
 
     pub fn sync_traverse_speed(&mut self) {
@@ -121,13 +342,19 @@ impl Winder2 {
         )
     }
 
+    /// Can wind capability check
     pub const fn can_wind(&self) -> bool {
+        // Check if tension arm is zeroed and traverse is homed
         self.tension_arm.zeroed
             && self.traverse_controller.is_homed()
             && !self.traverse_controller.is_going_home()
     }
 
+    /// Can go to inner limit capability check
     pub fn can_go_in(&self) -> bool {
+        // Check if traverse is homed, not in standby, not traversing
+        // Allow changing direction (even when going out)
+        // Disallow when homing is in progress
         self.traverse_controller.is_homed()
             && self.traverse_mode != TraverseMode::Standby
             && !self.traverse_controller.is_going_in()
@@ -136,7 +363,11 @@ impl Winder2 {
             && self.mode != Winder2Mode::Wind
     }
 
+    /// Can go to outer limit capability check
     pub fn can_go_out(&self) -> bool {
+        // Check if traverse is homed, not in standby, not traversing
+        // Allow changing direction (even when going in)
+        // Disallow when homing is in progress
         self.traverse_controller.is_homed()
             && self.traverse_mode != TraverseMode::Standby
             && !self.traverse_controller.is_going_out()
@@ -145,7 +376,10 @@ impl Winder2 {
             && self.mode != Winder2Mode::Wind
     }
 
+    /// Can go home capability check
     pub fn can_go_home(&self) -> bool {
+        // Check if not in standby, not traversing
+        // Allow going home even when going in or out
         self.traverse_mode != TraverseMode::Standby
             && !self.traverse_controller.is_going_home()
             && !self.traverse_controller.is_traversing()
@@ -153,98 +387,211 @@ impl Winder2 {
     }
 }
 
-/// Implement Mode transitions (without emit)
+impl Winder2 {}
+
+/// Implement Mode
 impl Winder2 {
+    fn set_mode(&mut self, mode: &Winder2Mode) {
+        let should_update = *mode != Winder2Mode::Wind || self.can_wind();
+
+        if should_update {
+            // all transitions are allowed
+            self.mode = mode.clone();
+
+            // Apply the mode changes to the spool and puller
+            self.set_spool_mode(mode);
+            self.set_puller_mode(mode);
+            self.set_traverse_mode(mode);
+        }
+        self.emit_state();
+    }
+
+    /// Apply the mode changes to the spool
+    ///
+    /// It contains a transition matrix for atomic changes.
+    /// It will set [`Self::spool_mode`]
     fn set_spool_mode(&mut self, mode: &Winder2Mode) {
+        // Convert to `Winder2Mode` to `SpoolMode`
         let mode: SpoolMode = mode.clone().into();
+
+        // Transition matrix
         match self.spool_mode {
             SpoolMode::Standby => match mode {
-                SpoolMode::Hold => self.spool.set_enabled(true),
+                SpoolMode::Standby => {}
+                SpoolMode::Hold => {
+                    // From [`SpoolMode::Standby`] to [`SpoolMode::Hold`]
+                    self.spool.set_enabled(true);
+                }
                 SpoolMode::Wind => {
                     self.spool.set_enabled(true);
+                    // self.spool_speed_controller.reset();
                     self.spool_speed_controller.set_enabled(true);
                 }
-                _ => {}
             },
             SpoolMode::Hold => match mode {
-                SpoolMode::Standby => self.spool.set_enabled(false),
-                SpoolMode::Wind => self.spool_speed_controller.set_enabled(true),
-                _ => {}
+                SpoolMode::Standby => {
+                    // From [`SpoolMode::Hold`] to [`SpoolMode::Standby`]
+                    self.spool.set_enabled(false);
+                }
+                SpoolMode::Hold => {}
+                SpoolMode::Wind => {
+                    // From [`SpoolMode::Hold`] to [`SpoolMode::Wind`]
+                    // self.spool_speed_controller.reset();
+                    self.spool_speed_controller.set_enabled(true);
+                }
             },
             SpoolMode::Wind => match mode {
                 SpoolMode::Standby => {
+                    // From [`SpoolMode::Wind`] to [`SpoolMode::Standby`]
                     self.spool.set_enabled(false);
                     self.spool_speed_controller.set_enabled(false);
                 }
-                SpoolMode::Hold => self.spool_speed_controller.set_enabled(false),
-                _ => {}
+                SpoolMode::Hold => {
+                    // From [`SpoolMode::Wind`] to [`SpoolMode::Hold`]
+                    self.spool_speed_controller.set_enabled(false);
+                }
+                SpoolMode::Wind => {}
             },
         }
+
+        // Update the internal state
         self.spool_mode = mode;
     }
 
+    /// Apply the mode changes to the spool
+    ///
+    /// It contains a transition matrix for atomic changes.
+    /// It will set [`Self::spool_mode`]
     fn set_traverse_mode(&mut self, mode: &Winder2Mode) {
+        // Convert to `Winder2Mode` to `TraverseMode`
         let mode: TraverseMode = mode.clone().into();
 
+        // If coming out of standby
         if self.traverse_mode == TraverseMode::Standby && mode != TraverseMode::Standby {
             self.traverse.set_enabled(true);
             self.traverse_controller.set_enabled(true);
         }
+
+        // If going into standby
         if mode == TraverseMode::Standby && self.traverse_mode != TraverseMode::Standby {
+            // If we are going into standby, we need to stop the traverse
             self.traverse.set_enabled(false);
             self.traverse_controller.set_enabled(false);
         }
 
+        // Transition matrix
         match self.traverse_mode {
             TraverseMode::Standby => match mode {
-                TraverseMode::Hold => self.traverse_controller.goto_home(),
-                TraverseMode::Traverse => self.traverse_controller.start_traversing(),
-                _ => {}
+                TraverseMode::Standby => {}
+                TraverseMode::Hold => {
+                    // From [`TraverseMode::Standby`] to [`TraverseMode::Hold`]
+                    self.traverse.set_enabled(true);
+                    self.traverse_controller.set_enabled(true);
+                    self.traverse_controller.goto_home();
+                }
+                TraverseMode::Traverse => {
+                    // From [`TraverseMode::Standby`] to [`TraverseMode::Wind`]
+                    self.traverse.set_enabled(true);
+                    self.traverse_controller.set_enabled(true);
+                    self.traverse_controller.start_traversing();
+                }
             },
             TraverseMode::Hold => match mode {
-                TraverseMode::Traverse => self.traverse_controller.start_traversing(),
-                _ => {}
+                TraverseMode::Standby => {
+                    // From [`TraverseMode::Hold`] to [`TraverseMode::Standby`]
+                    self.traverse.set_enabled(false);
+                    self.traverse_controller.set_enabled(false);
+                }
+                TraverseMode::Hold => {}
+                TraverseMode::Traverse => {
+                    // From [`TraverseMode::Hold`] to [`TraverseMode::Wind`]
+                    self.traverse_controller.start_traversing();
+                }
             },
             TraverseMode::Traverse => match mode {
-                TraverseMode::Hold => self.traverse_controller.goto_home(),
-                _ => {}
+                TraverseMode::Standby => {
+                    // From [`TraverseMode::Wind`] to [`TraverseMode::Standby`]
+                    self.traverse.set_enabled(false);
+                    self.traverse_controller.set_enabled(false);
+                }
+                TraverseMode::Hold => {
+                    // From [`TraverseMode::Wind`] to [`TraverseMode::Hold`]
+                    self.traverse_controller.goto_home();
+                }
+                TraverseMode::Traverse => {}
             },
         }
 
+        // Update the internal state
         self.traverse_mode = mode;
+        self.emit_state();
     }
 
+    /// Apply the mode changes to the puller
+    ///
+    /// It contains a transition matrix for atomic changes.
+    /// It will set [`Self::puller_mode`]
     fn set_puller_mode(&mut self, mode: &Winder2Mode) {
+        // Convert to `Winder2Mode` to `PullerMode`
         let mode: PullerMode = mode.clone().into();
+
+        // Transition matrix
         match self.puller_mode {
             PullerMode::Standby => match mode {
-                PullerMode::Hold => self.puller.set_enabled(true),
+                PullerMode::Standby => {}
+                PullerMode::Hold => {
+                    // From [`PullerMode::Standby`] to [`PullerMode::Hold`]
+                    self.puller.set_enabled(true);
+                }
                 PullerMode::Pull => {
+                    // From [`PullerMode::Standby`] to [`PullerMode::Pull`]
                     self.puller.set_enabled(true);
                     self.puller_speed_controller.set_enabled(true);
                 }
-                _ => {}
             },
             PullerMode::Hold => match mode {
-                PullerMode::Standby => self.puller.set_enabled(false),
-                PullerMode::Pull => self.puller_speed_controller.set_enabled(true),
-                _ => {}
+                PullerMode::Standby => {
+                    // From [`PullerMode::Hold`] to [`PullerMode::Standby`]
+                    self.puller.set_enabled(false);
+                }
+                PullerMode::Hold => {}
+                PullerMode::Pull => {
+                    // From [`PullerMode::Hold`] to [`PullerMode::Pull`]
+                    self.puller_speed_controller.set_enabled(true);
+                }
             },
             PullerMode::Pull => match mode {
                 PullerMode::Standby => {
+                    // From [`PullerMode::Pull`] to [`PullerMode::Standby`]
                     self.puller.set_enabled(false);
                     self.puller_speed_controller.set_enabled(false);
                 }
-                PullerMode::Hold => self.puller_speed_controller.set_enabled(false),
-                _ => {}
+                PullerMode::Hold => {
+                    // From [`PullerMode::Pull`] to [`PullerMode::Hold`]
+                    self.puller_speed_controller.set_enabled(false);
+                }
+                PullerMode::Pull => {}
             },
         }
+
+        // Update the internal state
         self.puller_mode = mode;
     }
 }
 
-/// Implement Spool calculations
+/// Implement Tension Arm
 impl Winder2 {
+    fn tension_arm_zero(&mut self) {
+        self.tension_arm.zero();
+        self.emit_live_values(); // For angle update
+        // For state update
+        self.emit_state();
+    }
+}
+
+/// Implement Spool
+impl Winder2 {
+    /// called by `act`
     pub fn sync_spool_speed(&mut self, t: Instant) {
         let angular_velocity = self.spool_speed_controller.update_speed(
             t,
@@ -257,6 +604,16 @@ impl Winder2 {
         let _ = self.spool.set_speed(steps_per_second);
     }
 
+    pub fn set_spool_automatic_required_meters(&mut self, meters: f64) {
+        self.spool_automatic_action.target_length = Length::new::<meter>(meters);
+        self.emit_state();
+    }
+
+    pub fn set_spool_automatic_mode(&mut self, mode: SpoolAutomaticActionMode) {
+        self.spool_automatic_action.mode = mode;
+        self.emit_state();
+    }
+
     pub fn stop_or_pull_spool(&mut self, now: Instant) {
         if matches!(
             self.spool_automatic_action.mode,
@@ -267,7 +624,8 @@ impl Winder2 {
         }
 
         match self.mode {
-            Winder2Mode::Pull | Winder2Mode::Wind => self.calculate_spool_auto_progress_(now),
+            Winder2Mode::Pull => self.calculate_spool_auto_progress_(now),
+            Winder2Mode::Wind => self.calculate_spool_auto_progress_(now),
             _ => {
                 self.spool_automatic_action.progress_last_check = now;
                 return;
@@ -276,15 +634,15 @@ impl Winder2 {
 
         if self.spool_automatic_action.progress >= self.spool_automatic_action.target_length {
             match self.spool_automatic_action.mode {
+                SpoolAutomaticActionMode::NoAction => (),
                 SpoolAutomaticActionMode::Pull => {
                     self.stop_or_pull_spool_reset(now);
-                    self.mode = Winder2Mode::Pull;
+                    self.set_mode(&Winder2Mode::Pull);
                 }
                 SpoolAutomaticActionMode::Hold => {
                     self.stop_or_pull_spool_reset(now);
-                    self.mode = Winder2Mode::Hold;
+                    self.set_mode(&Winder2Mode::Hold);
                 }
-                _ => {}
             }
         }
     }
@@ -295,10 +653,12 @@ impl Winder2 {
     }
 
     pub fn calculate_spool_auto_progress_(&mut self, now: Instant) {
+        // Calculate time elapsed since last progress check (in minutes)
         let dt = now
             .duration_since(self.spool_automatic_action.progress_last_check)
             .as_secs_f64();
 
+        // Calculate distance pulled during this time interval
         let meters_pulled_this_interval = Length::new::<meter>(
             self.puller_speed_controller
                 .last_speed
@@ -306,13 +666,15 @@ impl Winder2 {
                 * dt,
         );
 
+        // Update total meters pulled
         self.spool_automatic_action.progress += meters_pulled_this_interval;
         self.spool_automatic_action.progress_last_check = now;
     }
 }
 
-/// Implement Puller calculations
+/// Implement Puller
 impl Winder2 {
+    /// called by `act`
     pub fn sync_puller_speed(&mut self, t: Instant) {
         let angular_velocity = self.puller_speed_controller.calc_angular_velocity(t);
         let steps_per_second = self
@@ -321,10 +683,172 @@ impl Winder2 {
             .angular_velocity_to_steps(angular_velocity);
         let _ = self.puller.set_speed(steps_per_second);
     }
+
+    pub fn puller_set_regulation(&mut self, puller_regulation_mode: PullerRegulationMode) {
+        self.puller_speed_controller
+            .set_regulation_mode(puller_regulation_mode);
+        self.emit_state();
+    }
+
+    /// Set target speed in m/min
+    pub fn puller_set_target_speed(&mut self, target_speed: f64) {
+        // Convert m/min to velocity
+        let target_speed = Velocity::new::<meter_per_minute>(target_speed);
+        self.puller_speed_controller.set_target_speed(target_speed);
+        self.emit_state();
+    }
+
+    /// Set target diameter in mm
+    pub fn puller_set_target_diameter(&mut self, target_diameter: f64) {
+        // Convert m/min to velocity
+        let target_diameter = Length::new::<millimeter>(target_diameter);
+        self.puller_speed_controller
+            .set_target_diameter(target_diameter);
+        self.emit_state();
+    }
+
+    /// Set forward direction
+    pub fn puller_set_forward(&mut self, forward: bool) {
+        self.puller_speed_controller.set_forward(forward);
+        self.emit_state();
+    }
+
+    // Spool Speed Controller API methods
+    pub fn spool_set_regulation_mode(
+        &mut self,
+        regulation_mode: spool_speed_controller::SpoolSpeedControllerType,
+    ) {
+        self.spool_speed_controller.set_type(regulation_mode);
+        self.emit_state();
+    }
+
+    /// Set minimum speed for minmax mode in RPM
+    pub fn spool_set_minmax_min_speed(&mut self, min_speed_rpm: f64) {
+        let min_speed = uom::si::f64::AngularVelocity::new::<revolution_per_minute>(min_speed_rpm);
+        if let Err(e) = self.spool_speed_controller.set_minmax_min_speed(min_speed) {
+            tracing::error!("Failed to set spool min speed: {:?}", e);
+        }
+        self.emit_state();
+    }
+
+    /// Set maximum speed for minmax mode in RPM
+    pub fn spool_set_minmax_max_speed(&mut self, max_speed_rpm: f64) {
+        let max_speed = uom::si::f64::AngularVelocity::new::<revolution_per_minute>(max_speed_rpm);
+        if let Err(e) = self.spool_speed_controller.set_minmax_max_speed(max_speed) {
+            tracing::error!("Failed to set spool max speed: {:?}", e);
+        }
+        self.emit_state();
+    }
+
+    /// Set tension target for adaptive mode (0.0-1.0)
+    pub fn spool_set_adaptive_tension_target(&mut self, tension_target: f64) {
+        self.spool_speed_controller
+            .set_adaptive_tension_target(tension_target);
+        self.emit_state();
+    }
+
+    /// Set radius learning rate for adaptive mode
+    pub fn spool_set_adaptive_radius_learning_rate(&mut self, radius_learning_rate: f64) {
+        self.spool_speed_controller
+            .set_adaptive_radius_learning_rate(radius_learning_rate);
+        self.emit_state();
+    }
+
+    /// Set max speed multiplier for adaptive mode
+    pub fn spool_set_adaptive_max_speed_multiplier(&mut self, max_speed_multiplier: f64) {
+        self.spool_speed_controller
+            .set_adaptive_max_speed_multiplier(max_speed_multiplier);
+        self.emit_state();
+    }
+
+    /// Set acceleration factor for adaptive mode
+    pub fn spool_set_adaptive_acceleration_factor(&mut self, acceleration_factor: f64) {
+        self.spool_speed_controller
+            .set_adaptive_acceleration_factor(acceleration_factor);
+        self.emit_state();
+    }
+
+    /// Set deacceleration urgency multiplier for adaptive mode
+    pub fn spool_set_adaptive_deacceleration_urgency_multiplier(
+        &mut self,
+        deacceleration_urgency_multiplier: f64,
+    ) {
+        self.spool_speed_controller
+            .set_adaptive_deacceleration_urgency_multiplier(deacceleration_urgency_multiplier);
+        self.emit_state();
+    }
 }
 
-/// Implement machine connection helpers (without emit)
+/// implement machine connection
 impl Winder2 {
+    /// set connected buffer
+    pub fn set_connected_buffer(
+        &mut self,
+        machine_identification_unique: MachineIdentificationUnique,
+    ) {
+        if !matches!(
+            machine_identification_unique.machine_identification,
+            BufferV1::MACHINE_IDENTIFICATION
+        ) {
+            return;
+        }
+        let machine_manager_arc = match self.machine_manager.upgrade() {
+            Some(machine_manager_arc) => machine_manager_arc,
+            None => return,
+        };
+        let machine_manager_guard = block_on(machine_manager_arc.read());
+        let buffer_weak = machine_manager_guard.get_machine_weak(&machine_identification_unique);
+        let buffer_weak = match buffer_weak {
+            Some(buffer_weak) => buffer_weak,
+            None => return,
+        };
+        let buffer_strong = match buffer_weak.upgrade() {
+            Some(buffer_strong) => buffer_strong,
+            None => return,
+        };
+
+        let buffer: Arc<Mutex<BufferV1>> = block_on(downcast_machine::<BufferV1>(buffer_strong))
+            .expect("failed downcasting machine");
+
+        let machine = Arc::downgrade(&buffer);
+
+        self.connected_buffer = Some(ConnectedMachine {
+            machine_identification_unique,
+            machine,
+        });
+        self.emit_state();
+
+        self.reverse_connect();
+    }
+
+    /// disconnect buffer
+    pub fn disconnect_buffer(
+        &mut self,
+        machine_identification_unique: MachineIdentificationUnique,
+    ) {
+        if !matches!(
+            machine_identification_unique.machine_identification,
+            BufferV1::MACHINE_IDENTIFICATION
+        ) {
+            return;
+        }
+        if let Some(connected) = &self.connected_buffer {
+            if let Some(buffer_arc) = connected.machine.upgrade() {
+                let future = async move {
+                    let mut buffer = buffer_arc.lock().await;
+                    if buffer.connected_winder.is_some() {
+                        buffer.connected_winder = None;
+                        buffer.emit_state();
+                    }
+                };
+                smol::spawn(future).detach();
+            }
+        }
+        self.connected_buffer = None;
+        self.emit_state();
+    }
+
+    /// initiate connection from buffer to winder
     pub fn reverse_connect(&mut self) {
         let machine_identification_unique = self.machine_identification_unique.clone();
         if let Some(connected) = &self.connected_buffer {
@@ -341,7 +865,6 @@ impl Winder2 {
     }
 }
 
-// Enums
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum Winder2Mode {
     Standby,
@@ -407,5 +930,54 @@ impl From<Winder2Mode> for PullerMode {
 impl std::fmt::Display for Winder2 {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         write!(f, "Winder2")
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use uom::si::{f64::Length, length::millimeter};
+
+    #[test]
+    fn test_validate_traverse_limits() {
+        // Test case 1: Valid limits with exactly 1.0mm difference (should pass)
+        let inner = Length::new::<millimeter>(15.0);
+        let outer = Length::new::<millimeter>(16.0);
+        assert!(Winder2::validate_traverse_limits(inner, outer));
+
+        // Test case 2: Invalid limits with exactly 0.9mm difference (should fail)
+        let inner = Length::new::<millimeter>(15.0);
+        let outer = Length::new::<millimeter>(15.9);
+        assert!(!Winder2::validate_traverse_limits(inner, outer));
+
+        // Test case 3: Invalid limits with less than 0.9mm difference (should fail)
+        let inner = Length::new::<millimeter>(15.0);
+        let outer = Length::new::<millimeter>(15.5);
+        assert!(!Winder2::validate_traverse_limits(inner, outer));
+
+        // Test case 4: Invalid limits where inner equals outer (should fail)
+        let inner = Length::new::<millimeter>(20.0);
+        let outer = Length::new::<millimeter>(20.0);
+        assert!(!Winder2::validate_traverse_limits(inner, outer));
+
+        // Test case 5: Invalid limits where inner is greater than outer (should fail)
+        let inner = Length::new::<millimeter>(25.0);
+        let outer = Length::new::<millimeter>(20.0);
+        assert!(!Winder2::validate_traverse_limits(inner, outer));
+
+        // Test case 6: Valid limits with large difference (should pass)
+        let inner = Length::new::<millimeter>(10.0);
+        let outer = Length::new::<millimeter>(80.0);
+        assert!(Winder2::validate_traverse_limits(inner, outer));
+
+        // Test case 7: Edge case - exactly 0.91mm difference (should pass)
+        let inner = Length::new::<millimeter>(15.0);
+        let outer = Length::new::<millimeter>(15.91);
+        assert!(Winder2::validate_traverse_limits(inner, outer));
+
+        // Test case 8: Edge case - exactly 0.89mm difference (should fail)
+        let inner = Length::new::<millimeter>(15.0);
+        let outer = Length::new::<millimeter>(15.89);
+        assert!(!Winder2::validate_traverse_limits(inner, outer));
     }
 }
